move node metadata and distributed locking to OHazelcastPlugin
